+++ sdk-rust/src/opentelemetry.rs
@@ -22,7 +22,7 @@ pub struct LMSpan {
     max_tokens: Option<u32>,
     temperature: Option<f64>,
     top_p: Option<f64>,
-    top_k: Option<f64>,
+    top_k: Option<i32>,
     presence_penalty: Option<f64>,
     frequency_penalty: Option<f64>,
     seed: Option<i64>,
@@ -135,8 +135,10 @@ impl LMSpan {
                 .set_attribute(KeyValue::new("gen_ai.request.top_p", top_p));
         }
         if let Some(top_k) = self.top_k {
-            self.span
-                .set_attribute(KeyValue::new("gen_ai.request.top_k", top_k));
+            self.span.set_attribute(KeyValue::new(
+                "gen_ai.request.top_k",
+                i64::from(top_k),
+            ));
         }
         if let Some(presence_penalty) = self.presence_penalty {
             self.span.set_attribute(KeyValue::new(
